ml/train/requirements.txt
torch
torchvision
pillow
numpy
opencv-python
tqdm
scikit-learn
